Create ICyrelaCrawler e Handler

+++ BoraMorar/Indices/Repository/IIndiceAplicadoRepository.cs
@@ -2,5 +2,5 @@
 
 public interface IIndiceAplicadoRepository : IAddRepository<IndiceAplicado>
 {
-    Task<IndiceAplicado?> FindAsync(int id);
+    Task<IndiceAplicado?> FindAsync(long id);
 }
